Remove KernelCenterer
Old version of sklearn kernelcenterer accepted nonsquare matrices, this is no longer the case with the new package. Looking through the docs for KernelCenterer, I don't see the need for it as we are already centering our features with StandardScalar. As well, since we are preprocessing both the training and testing data in the same way, I am not too worried.

I have run both data sets (mnist, wisc).  mnist results are pretty much identical as the old code (it's heuristic and is off by at most 1%), while wisc results are exactly the same in training but have improved in testing