Add --cross-validation arg and do single run by default
Instead of defaulting to a cross-validation parameter sweep over
lambda, use a single lambda value by default.  This speeds up the runs
roughly by a factor of 10.  The parameter sweep can still be done by
providing the --cross-validation option.  When cross validation is
done, a message is added to the screen output indicating that the run
may take several minutes.

README has also been updated to reflect the fact that cross-validation
is not done by default.